+++ internal/provider/resource_rediscloud_subscription.go
@@ -67,6 +67,14 @@ func resourceRedisCloudSubscription() *schema.Resource {
 				Type:        schema.TypeString,
 				Optional:    true,
 			},
+			"payment_method": {
+				Description:      "Payment method for the requested subscription. If credit card is specified, the payment method Id must be defined.",
+				Type:             schema.TypeString,
+				ForceNew:         true,
+				ValidateDiagFunc: validateDiagFunc(validation.StringMatch(regexp.MustCompile("^(credit-card|marketplace)$"), "must be 'credit-card' or 'marketplace'")),
+				Optional:         true,
+				Default:          "credit-card",
+			},
 			"payment_method_id": {
 				Computed:         true,
 				Description:      "A valid payment method pre-defined in the current account",
@@ -356,10 +364,9 @@ func resourceRedisCloudSubscription() *schema.Resource {
 						},
 						"module": {
 							Description: "A module object",
-							Type:        schema.TypeList,
+							Type:        schema.TypeSet,
 							Optional:    true,
 							MinItems:    1,
-							MaxItems:    1,
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
 									"name": {
@@ -418,6 +425,7 @@ func resourceRedisCloudSubscriptionCreate(ctx context.Context, d *schema.Resourc
 	// Create Subscription
 	name := d.Get("name").(string)
 
+	paymentMethod := d.Get("payment_method").(string)
 	paymentMethodID, err := readPaymentMethodID(d)
 	if err != nil {
 		return diag.FromErr(err)
@@ -429,6 +437,7 @@ func resourceRedisCloudSubscriptionCreate(ctx context.Context, d *schema.Resourc
 		Name:            redis.String(name),
 		DryRun:          redis.Bool(false),
 		PaymentMethodID: paymentMethodID,
+		PaymentMethod:   redis.String(paymentMethod),
 		MemoryStorage:   redis.String(memoryStorage),
 		CloudProviders:  providers,
 		Databases:       dbs,
@@ -494,6 +503,9 @@ func resourceRedisCloudSubscriptionRead(ctx context.Context, d *schema.ResourceD
 		if err := d.Set("payment_method_id", paymentMethodID); err != nil {
 			return diag.FromErr(err)
 		}
+	}
+	if err := d.Set("payment_method", redis.StringValue(subscription.PaymentMethod)); err != nil {
+		return diag.FromErr(err)
 	}
 	if err := d.Set("memory_storage", redis.StringValue(subscription.MemoryStorage)); err != nil {
 		return diag.FromErr(err)
@@ -592,6 +604,7 @@ func resourceRedisCloudSubscriptionUpdate(ctx context.Context, d *schema.Resourc
 		} else {
 			// this is not a new resource, so these databases really do new to be created
 			for _, db := range addition {
+			// This loop with addition is triggered when another database is added to the subscription.
 				request := buildCreateDatabase(db)
 				id, err := api.client.Database.Create(ctx, subId, request)
 				if err != nil {
@@ -776,8 +789,8 @@ func buildSubscriptionCreateDatabases(databases interface{}) []*subscriptions.Cr
 		averageItemSizeInBytes := databaseMap["average_item_size_in_bytes"].(int)
 
 		createModules := make([]*subscriptions.CreateModules, 0)
-		modules := databaseMap["module"]
-		for _, module := range modules.([]interface{}) {
+		modules := databaseMap["module"].(*schema.Set)
+		for _, module := range modules.List() {
 			moduleMap := module.(map[string]interface{})
 
 			modName := moduleMap["name"].(string)
@@ -826,8 +839,8 @@ func buildCreateDatabase(db map[string]interface{}) databases.CreateDatabase {
 	}
 
 	createModules := make([]*databases.CreateModule, 0)
-	module := db["module"]
-	for _, module := range module.([]interface{}) {
+	module := db["module"].(*schema.Set)
+	for _, module := range module.List() {
 		moduleMap := module.(map[string]interface{})
 
 		modName := moduleMap["name"].(string)
@@ -1304,30 +1317,43 @@ func getDatabaseNameIdMap(ctx context.Context, subId int, client *apiClient) (ma
 	return ret, nil
 }
 
+// diff: Checks the difference between two Sets based on their hash keys and check if they were modified by generating
+//       a hash based on their attributes.
 func diff(oldSet *schema.Set, newSet *schema.Set, hashKey func(interface{}) string) ([]map[string]interface{}, []map[string]interface{}, []map[string]interface{}) {
-	oldMap := map[string]map[string]interface{}{}
-	newMap := map[string]map[string]interface{}{}
+
+	oldHashedMap := map[string]*hashedSet{}
+	newHashedMap := map[string]*hashedSet{}
 
 	for _, v := range oldSet.List() {
-		oldMap[hashKey(v)] = v.(map[string]interface{})
+		h := hashedSet{}
+		oldHashedMap[hashKey(v)] = h.init(oldSet, v)
 	}
 	for _, v := range newSet.List() {
-		newMap[hashKey(v)] = v.(map[string]interface{})
+		h := hashedSet{}
+		newHashedMap[hashKey(v)] = h.init(newSet, v)
 	}
 
 	var addition, existing, deletion []map[string]interface{}
 
-	for k, v := range newMap {
-		if _, ok := oldMap[k]; ok {
-			existing = append(existing, v)
+	for k, newVal := range newHashedMap {
+		// Check if we're updating an existing block.
+		if oldVal, ok := oldHashedMap[k]; ok {
+			// The hashes are the same - this block has NOT been changed.
+			if oldVal.hash == newVal.hash {
+				continue
+			}
+			// The hashes are different - this block was modified.
+			existing = append(existing, newVal.m)
+		// This block was recently added.
 		} else {
-			addition = append(addition, v)
+			addition = append(addition, newVal.m)
 		}
 	}
 
-	for k, v := range oldMap {
-		if _, ok := newMap[k]; !ok {
-			deletion = append(deletion, v)
+	for k, oldVal := range oldHashedMap {
+		// This block was deleted.
+		if _, ok := newHashedMap[k]; !ok {
+			deletion = append(deletion, oldVal.m)
 		}
 	}