+++ internal/provider/resource_rediscloud_subscription_test.go
@@ -18,6 +18,9 @@ import (
 var contractFlag = flag.Bool("contract", false,
 	"Add this flag '-contract' to run tests for contract associated accounts")
 
+var marketplaceFlag = flag.Bool("marketplace", false,
+	"Add this flag '-marketplace' to run tests for marketplace associated accounts")
+
 func TestAccResourceRedisCloudSubscription_createWithDatabase(t *testing.T) {
 
 	name := acctest.RandomWithPrefix(testResourcePrefix)
@@ -278,6 +281,47 @@ func TestAccResourceRedisCloudSubscription_AddAdditionalDatabaseWithModule(t *te
 	})
 }
 
+// Steps:
+// - Create a subscription with a multi-module db.
+// - Add another multi-module db to the subscription.
+func TestAccResourceRedisCloudSubscription_AddDatabasesWithMultiModules(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Requires manual execution over CI execution")
+	}
+
+	name := acctest.RandomWithPrefix(testResourcePrefix)
+	password := acctest.RandString(20)
+	resourceName := "rediscloud_subscription.example"
+	testCloudAccountName := os.Getenv("AWS_TEST_CLOUD_ACCOUNT_NAME")
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { testAccPreCheck(t); testAccAwsPreExistingCloudAccountPreCheck(t) },
+		ProviderFactories: providerFactories,
+		CheckDestroy:      testAccCheckSubscriptionDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(testAccResourceRedisCloudSubscriptionDbWithMultipleModules, testCloudAccountName, name, password),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "database.0.module.#", "2"),
+					resource.TestCheckResourceAttr(resourceName, "database.0.module.0.name", "RedisBloom"),
+					resource.TestCheckResourceAttr(resourceName, "database.0.module.1.name", "RedisJSON"),
+				),
+			},
+			{
+				Config: fmt.Sprintf(testAccResourceRedisCloudSubscriptionTwoDbsWithMultipleModules, testCloudAccountName, name, password, password),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "database.0.module.#", "2"),
+					resource.TestCheckResourceAttr(resourceName, "database.0.module.0.name", "RedisBloom"),
+					resource.TestCheckResourceAttr(resourceName, "database.0.module.1.name", "RedisJSON"),
+					resource.TestCheckResourceAttr(resourceName, "database.1.module.#", "2"),
+					resource.TestCheckResourceAttr(resourceName, "database.1.module.0.name", "RedisBloom"),
+					resource.TestCheckResourceAttr(resourceName, "database.1.module.1.name", "RedisJSON"),
+				),
+			},
+		},
+	})
+}
+
 func TestAccResourceRedisCloudSubscription_AddManageDatabaseReplication(t *testing.T) {
 
 	if testing.Short() {
@@ -363,6 +407,47 @@ func TestAccResourceRedisCloudSubscription_createUpdateContractPayment(t *testin
 				Config: fmt.Sprintf(testAccResourceRedisCloudSubscriptionContractPayment, testCloudAccountName, updatedName, 1, password),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttrSet(resourceName, "payment_method_id"),
+					resource.TestCheckResourceAttr(resourceName, "name", updatedName),
+				),
+			},
+		},
+	})
+}
+
+func TestAccResourceRedisCloudSubscription_createUpdateMarketplacePayment(t *testing.T) {
+
+	if !*marketplaceFlag {
+		t.Skip("The '-marketplace' parameter wasn't provided in the test command.")
+	}
+
+	name := acctest.RandomWithPrefix(testResourcePrefix)
+	updatedName := fmt.Sprintf("%v-updatedName", name)
+	password := acctest.RandString(20)
+	resourceName := "rediscloud_subscription.example"
+	testCloudAccountName := os.Getenv("AWS_TEST_CLOUD_ACCOUNT_NAME")
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { testAccPreCheck(t); testAccAwsPreExistingCloudAccountPreCheck(t) },
+		ProviderFactories: providerFactories,
+		CheckDestroy:      testAccCheckSubscriptionDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(testAccResourceRedisCloudSubscriptionMarketplacePayment, testCloudAccountName, name, 1, password),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "name", name),
+					resource.TestCheckResourceAttr(resourceName, "cloud_provider.0.provider", "AWS"),
+					resource.TestCheckResourceAttr(resourceName, "cloud_provider.0.region.0.preferred_availability_zones.#", "1"),
+					resource.TestCheckResourceAttrSet(resourceName, "cloud_provider.0.region.0.networks.0.networking_subnet_id"),
+					resource.TestCheckResourceAttr(resourceName, "database.#", "1"),
+					resource.TestMatchResourceAttr(resourceName, "database.0.db_id", regexp.MustCompile("^[1-9][0-9]*$")),
+					resource.TestCheckResourceAttrSet(resourceName, "database.0.password"),
+					resource.TestCheckResourceAttr(resourceName, "database.0.name", "tf-database"),
+					resource.TestCheckResourceAttr(resourceName, "database.0.memory_limit_in_gb", "1"),
+				),
+			},
+			{
+				Config: fmt.Sprintf(testAccResourceRedisCloudSubscriptionMarketplacePayment, testCloudAccountName, updatedName, 1, password),
+				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "name", updatedName),
 				),
 			},
@@ -569,6 +654,133 @@ resource "rediscloud_subscription" "example" {
 }
 `
 
+const testAccResourceRedisCloudSubscriptionDbWithMultipleModules = `
+data "rediscloud_payment_method" "card" {
+  card_type = "Visa"
+}
+
+data "rediscloud_cloud_account" "account" {
+  exclude_internal_account = true
+  provider_type = "AWS"
+  name = "%s"
+}
+
+resource "rediscloud_subscription" "example" {
+
+  name = "%s"
+  payment_method_id = data.rediscloud_payment_method.card.id
+  memory_storage = "ram"
+
+  allowlist {
+    cidrs = ["192.168.0.0/16"]
+  }
+
+  cloud_provider {
+    provider = data.rediscloud_cloud_account.account.provider_type
+    cloud_account_id = data.rediscloud_cloud_account.account.id
+    region {
+      region = "eu-west-1"
+      networking_deployment_cidr = "10.0.0.0/24"
+      preferred_availability_zones = ["eu-west-1a"]
+    }
+  }
+
+  database {
+    name = "tf-database-0"
+	protocol = "redis"
+    memory_limit_in_gb = 1
+    support_oss_cluster_api = true
+    data_persistence = "none"
+    replication = false
+    throughput_measurement_by = "operations-per-second"
+    password = "%s"
+    throughput_measurement_value = 10000
+    source_ips = ["10.0.0.0/8"]
+
+    dynamic "module" {
+      for_each = ["RedisJSON", "RedisBloom"]
+      content {
+        name = module.value
+      }
+    }
+  }
+}
+`
+
+const testAccResourceRedisCloudSubscriptionTwoDbsWithMultipleModules = `
+data "rediscloud_payment_method" "card" {
+  card_type = "Visa"
+}
+
+data "rediscloud_cloud_account" "account" {
+  exclude_internal_account = true
+  provider_type = "AWS"
+  name = "%s"
+}
+
+resource "rediscloud_subscription" "example" {
+
+  name = "%s"
+  payment_method_id = data.rediscloud_payment_method.card.id
+  memory_storage = "ram"
+
+  allowlist {
+    cidrs = ["192.168.0.0/16"]
+  }
+
+  cloud_provider {
+    provider = data.rediscloud_cloud_account.account.provider_type
+    cloud_account_id = data.rediscloud_cloud_account.account.id
+    region {
+      region = "eu-west-1"
+      networking_deployment_cidr = "10.0.0.0/24"
+      preferred_availability_zones = ["eu-west-1a"]
+    }
+  }
+
+  database {
+    name = "tf-database-0"
+	protocol = "redis"
+    memory_limit_in_gb = 1
+    support_oss_cluster_api = true
+    data_persistence = "none"
+    replication = false
+    throughput_measurement_by = "operations-per-second"
+    password = "%s"
+    throughput_measurement_value = 10000
+    source_ips = ["10.0.0.0/8"]
+
+    dynamic "module" {
+      for_each = ["RedisJSON", "RedisBloom"]
+      content {
+        name = module.value
+      }
+    }
+  }
+  
+  database {
+    name = "tf-database-1"
+	protocol = "redis"
+    memory_limit_in_gb = 1
+    support_oss_cluster_api = true
+    data_persistence = "none"
+    replication = false
+    throughput_measurement_by = "operations-per-second"
+    password = "%s"
+    throughput_measurement_value = 10000
+    source_ips = ["10.0.0.0/8"]
+
+    dynamic "module" {
+      for_each = ["RedisJSON", "RedisBloom"]
+      content {
+        name = module.value
+      }
+    }
+  }
+  
+}
+`
+
 const testAccResourceRedisCloudSubscriptionsWithReplicaDB = `
 locals {
   test_cloud_account_name = "%s"
@@ -762,6 +974,49 @@ resource "rediscloud_subscription" "example" {
     support_oss_cluster_api = true
     data_persistence = "none"
 	data_eviction = "volatile-lru"
+    replication = false
+    throughput_measurement_by = "operations-per-second"
+    password = "%s"
+    throughput_measurement_value = 10000
+    source_ips = ["10.0.0.0/8"]
+  }
+}
+`
+
+const testAccResourceRedisCloudSubscriptionMarketplacePayment = `
+
+data "rediscloud_cloud_account" "account" {
+  exclude_internal_account = true
+  provider_type = "AWS" 
+  name = "%s"
+}
+
+resource "rediscloud_subscription" "example" {
+
+  name = "%s"
+  memory_storage = "ram"
+  payment_method = "marketplace"
+
+  allowlist {
+    cidrs = ["192.168.0.0/16"]
+  }
+
+  cloud_provider {
+    provider = data.rediscloud_cloud_account.account.provider_type
+    cloud_account_id = data.rediscloud_cloud_account.account.id
+    region {
+      region = "eu-west-1"
+      networking_deployment_cidr = "10.0.0.0/24"
+      preferred_availability_zones = ["eu-west-1a"]
+    }
+  }
+
+  database {
+    name = "tf-database"
+    protocol = "redis"
+    memory_limit_in_gb = %d
+    support_oss_cluster_api = true
+    data_persistence = "none"
     replication = false
     throughput_measurement_by = "operations-per-second"
     password = "%s"